requirements.txt
pillow==10.0.0
docopt==0.6.2
pytest==7.4.0
